lambdas/nodes/image_thumbnail/requirements.txt
pillow==11.0.0
cairosvg
aws-xray-sdk
aws-lambda-powertools